These are the recorded outputs from the system.
These are the recorded outputs from the system depending on the contextual user inputs.
Thus are the inputs received are converted into jason for machine level understanding and then the inputs are passed through the Voice Interface system that invokes the users input to call the lamda code. And thus generating outputs for all kinds of possible utterences, calling the particular intent for the user inputs and thus the out is generated. The generated out put is then converted into the speech context and thus the jason put is received into normal text or speech data.